refactor typed module
